requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
attrs==19.3.0
certifi==2019.9.11
chardet==3.0.4
discord.py==1.2.5
docopt==0.6.2
idna==2.8
monotonic==1.5
multidict==4.6.1
python-valve==0.2.1
requests==2.22.0
six==1.13.0
urllib3==1.25.7
websockets==8.1
yarl==1.3.0
